Migrate 18CO to new structure
Migrate 18CO to new structure
