--- lib/engine/step/g_18_co/track.rb
@@ -1,32 +1,0 @@
-# frozen_string_literal: true
-
-require_relative '../track'
-require_relative 'tracker'
-
-module Engine
-  module Step
-    module G18CO
-      class Track < Track
-        include Tracker
-
-        def process_lay_tile(action)
-          lay_tile_action(action)
-          clear_upgrade_icon(action.hex.tile)
-          collect_mines(action.entity, action.hex)
-          migrate_reservations(action.hex.tile)
-          # place_home_token(action.hex.tile, pending_token[:token]) if pending_token
-
-          pass! unless can_lay_tile?(action.entity)
-        end
-
-        def available_hex(entity, hex)
-          if pending_token(entity)
-            return hex == @game.hex_by_id(entity.coordinates) ? hex.tile.exits : nil
-          end
-
-          super
-        end
-      end
-    end
-  end
-end
